Analysis Methods

Analysis Methods

GraphEvolve.compute_Δ_method_1! — Method.

compute_Δ_method_1!(g::AbstractGraph)

Determines t₀ and t₁ for the given evolved g

Arguments

g : An evolved instance of type AbstractGraph

Extra Information

t₀: The last step in the evolution process satisfying g.observables.largest_cluster_size < sqrt(g.n)

t₁: The first step in the evolution process satisfying g.observables.largest_cluster_size > 0.5g.n

Returns

None, updates g in-place

source

GraphEvolve.compute_delta! — Method.

compute_delta!(g::AbstractGraph)

Determines t₀ and t₁ for the given evolved g

Arguments

g : An evolved instance of type AbstractGraph

Extra Information

t₀: The step in the evolution process where the heterogeneity peaks

t₁: The step in the evolution process where the size of the largest cluster increases the most

Returns

None, updates g in-place

source

GraphEvolve.finalize_observables! — Method.

finalize_observables!(g::AbstractGraph)

Finalize the observables for the given evolved g

Arguments

g : An evolved instance of type AbstractGraph

Returns

None, updates g in-place

source