Cluster Methods

Cluster Methods

GraphEvolve.get_cluster — Method.

get_cluster(g::AbstractGraph, node::Int)

Determines the cluster in g which node is a member of

Arguments

g : An instance of type AbstractGraph

node: Node in keys(g.cluster_ids)

Returns

Set of nodes representing the cluster which node is a member of

source

GraphEvolve.get_largest_cluster_size — Method.

get_largest_cluster_size(g::AbstractGraph)

Determines the size of the largest cluster in g

Arguments

g: An instance of type AbstractGraph

Returns

Integer representing the number of nodes in the largest cluster in g

source

GraphEvolve.get_avg_cluster_size — Method.

get_avg_cluster_size(g::AbstractGraph)

Determines the average cluster size in g

Arguments

g: An instance of type AbstractGraph

Returns

Float64 representing the average cluster size in g

source

GraphEvolve.get_largest_clusters — Method.

get_largest_clusters(g::AbstractGraph, n_clusters::Int)

Determines the n_clusters largest clusters in 'g'

Arguments

g : An instance of type AbstractGraph

n_clusters: The number of largest clusters to return

Returns

Sorted (descending) array of the n_clusters largest clusters in g

source

GraphEvolve.update_clusters! — Method.

update_clusters!(g::AbstractGraph, edge::Tuple)

Updates g with the newly merged cluster and the largest cluster size Arguments

g : An instance of type AbstractGraph

edge: Edge added to g at step g.t

Returns

None, updates g in-place

source

GraphEvolve.update_cluster_sizes! — Method.

update_cluster_sizes!(g::AbstractGraph, larger_cluster_id::Int, smaller_cluster_id::Int)

Updates the cluster size distribution dictionary

Arguments

g : An instance of type AbstractGraph

larger_cluster_id : Cluster ID of the larger cluster

smaller_cluster_id: Cluster ID of the smaller cluster

Returns

None, updates g in-place

source

GraphEvolve.update_cluster_ids! — Method.

update_cluster_ids!(g::AbstractGraph, larger_cluster_id::Int, smaller_cluster_id::Int)

Updates the cluster IDs of the nodes in the smaller cluster to that of the larger cluster it is being merged into

Arguments

g : An instance of type AbstractGraph

larger_cluster_id : Cluster ID of the larger cluster

smaller_cluster_id: Cluster ID of the smaller cluster

Returns

None, updates g in-place

source

GraphEvolve.merge_clusters! — Method.

merge_clusters!(g::AbstractGraph, larger_cluster_id::Int, smaller_cluster_id::Int)

Mergers the smaller cluster into the larger cluster in-place

Arguments

g : An instance of type AbstractGraph

larger_cluster_id : Cluster ID of the larger cluster

smaller_cluster_id: Cluster ID of the smaller cluster

Returns

None, updates g in-place

source

GraphEvolve.update_observables! — Method.

update_observables!(g::AbstractGraph, larger_cluster_id::Int, smaller_cluster_id::Int)

Updates the largest cluster size, average cluster size, and cluster heterogeneity

Arguments

g : An instance of type AbstractGraph

larger_cluster_id : Cluster ID of the larger cluster

smaller_cluster_id: Cluster ID of the smaller cluster

Returns

None, updates g in-place

source