Evolution Methods

Evolution Methods

GraphEvolve.erdos_renyi! — Method.

erdos_renyi!(g::AbstractGraph, n_steps::Int)

Erdos-Renyi style graph evolution, adds edges randomly at each step

Arguments

g : An instance of type AbstractGraph

n_steps: Number of edges to add to the AbstractGraph

Returns

nothing, updates g in-place

source

GraphEvolve.bohman_frieze! — Method.

bohman_frieze!(g::AbstractGraph, n_steps::Int; K::Int=2)

Achlioptas process, implementation of Bohman-Frieze bounded size rule

Arguments

g : An instance of type AbstractGraph

n_steps: Number of edges to add to the AbstractGraph

Keyword Arguments

K : Bounded size of clusters upon which to determine edge acceptance (default = 1)

Returns

nothing, updates g in-place

source

GraphEvolve.product_rule! — Method.

product_rule!(g::AbstractGraph, n_steps::Int)

Achlioptas process, implementation of the product rule

Arguments

g : An instance of type AbstractGraph

n_steps: Number of edges to add to the AbstractGraph

Returns

nothing, updates g in-place

source

GraphEvolve.stochastic_edge_acceptance! — Method.

stochastic_edge_acceptance!(g::AbstractGraph, n_steps::Int)

Achlioptas process, a probability based rule for accepting edges

Arguments

g : An instance of type AbstractGraph

n_steps: Number of edges to add to the AbstractGraph

q : Minimum probability that edge₁ is accepted

Returns

nothing, updates g in-place

source