Analysis Methods

GraphEvolve.compute_delta! — Method

compute_delta!(g::AbstractGraph)

Determines t₀ and t₁ for the given evolved g

Arguments

g : An evolved instance of type AbstractGraph

Extra Information

t₀: The step in the evolution process where the heterogeneity peaks

t₁: The step in the evolution process where the size of the largest cluster increases the most

Returns

nothing, updates g in-place

GraphEvolve.finalize_observables! — Method

finalize_observables!(g::AbstractGraph)

Finalize the observables for the given evolved g

Arguments

g : An evolved instance of type AbstractGraph

Returns

nothing, updates g in-place