Edge Methods

GraphEvolve.choose_edge — Method

choose_edge(g::Network)

Randomly selects an inactive edge in g

Arguments

g : An instance of type Network

Returns

edge: A two-tuple of integers representing an inactive edge in g

GraphEvolve.choose_edge — Method

choose_edge(g::Lattice2D)

Randomly selects an inactive edge in g

Arguments

g : An instance of type Lattice2D

Returns

edge: A two-tuples of integers representing an inactive edge in g

GraphEvolve.choose_edge — Method

choose_edge(g::Lattice3D)

Randomly selects an inactive edge in g

Arguments

g : An instance of type Lattice3D

Returns

edge: A two-tuple of integers representing an inactive edge in g

GraphEvolve.choose_edge — Method

choose_edge(g::AbstractGraph, edge₁::Tuple)

Randomly selects an inactive edge in g that is not equal to edge₁

Arguments

g : An instance of type AbstractGraph

edge₁: A two-tuple of integers representing an inactive edge in g

Returns

edge₂: A two-tuple of integers representing an inactive edge in g not equal to edge₁

GraphEvolve.choose_edge — Method

choose_edge(g::AbstractGraph, edges::Array{Tuple{Int, Int}, 1})

Randomly selects an inactive edge in g that is not equal to edge₁

Arguments

g : An instance of type AbstractGraph

edges: An array of two-tuples of integers representing inactive edges in g

Returns

edge: A two-tuple of integers representing an inactive edge in g not in edges

GraphEvolve.add_edge! — Method

add_edge!(g::AbstractGraph, edge::Tuple)

Adds an edge to g

Arguments

g : An instance of type AbstractGraph

edge: A two-tuple of integers representing the edge to be added to g

Returns

None, updates g in-place

GraphEvolve.nearest_neighbors — Method

nearest_neighbors(g::Lattice2D, node::Tuple{Int, Int})

Determines the nearest neighbors of node

Arguments

g : An instance of type Lattice2D

node : A two-tuple of integers representing a node in g

Returns

neighbors: A four-tuple of two-tuples of integers representing the cartesian indices of the (up, down, left, right) neighbors

GraphEvolve.nearest_neighbors — Method

nearest_neighbors(g::Lattice3D, node::Tuple{Int, Int, Int})

Determines the nearest neighbors of node

Arguments

g : An instance of type Lattice3D

node : A three-tuple of integers representing a node in g

Returns

neighbors: A four-tuple of three-tuples of integers representing the cartesian indices of the (up, down, left, right, front, back) neighbors

GraphEvolve.cart_to_lin — Method

cart_to_lin(cart::Tuple, L::Int)

Converts d-dimensional Cartesian index (d-tuple) to linear index

Arguments

cart : d-Tuple representing Cartesian index, d ∈ {2, 3}

L : Linear lattice size

Returns

lin : Linear index in the lattice corresponding to cart