Evolution Methods

GraphEvolve.erdos_renyi! — Method

erdos_renyi!(g::AbstractGraph, n_steps::Integer)

Erdos-Renyi style graph evolution, adds edges randomly at each step

Arguments

g : An instance of type AbstractGraph

n_steps: Number of edges to add to the AbstractGraph

Returns

g, updates g in-place

GraphEvolve.bohman_frieze! — Method

bohman_frieze!(g::AbstractGraph, n_steps::Integer; K::Integer=2)

Achlioptas process, implementation of Bohman-Frieze bounded size rule

Arguments

g : An instance of type AbstractGraph

n_steps: Number of edges to add to the AbstractGraph

Keyword Arguments

K : Bounded size of clusters upon which to determine edge acceptance (default = 2)

Returns

g, updates g in-place

GraphEvolve.product_rule! — Method

product_rule!(g::AbstractGraph, n_steps::Integer)

Achlioptas process, implementation of the product rule

Arguments

g : An instance of type AbstractGraph

n_steps: Number of edges to add to the AbstractGraph

Returns

g, updates g in-place

GraphEvolve.stochastic_edge_acceptance! — Method

stochastic_edge_acceptance!(g::AbstractGraph, n_steps::Integer)

Achlioptas process, a probability based model for accepting edges

Arguments

g : An instance of type AbstractGraph

n_steps: Number of edges to add to the AbstractGraph

Returns

g, updates g in-place

GraphEvolve.stochastic_edge_acceptance! — Method

stochastic_edge_acceptance!(g::AbstractGraph, n_steps::Integer, q::Integer)

q-edge Achlioptas process, a probability based model for accepting edges

Arguments

g : An instance of type AbstractGraph

n_steps: Number of edges to add to the AbstractGraph

q : Number of edges to evaluate

Returns

g, updates g in-place